refactor: separate mediumCard component into general component and item component
