+++ src/components/Cards/MediumCard/styles.ts
@@ -3,6 +3,8 @@ import styled from 'styled-components';
 export const Container = styled.div`
   width: 100%;
   height: 450px;
+
+  margin-bottom: 40px;
 `;
 
 export const Content = styled.div`
@@ -15,8 +17,6 @@ export const Content = styled.div`
 
   display: flex;
   align-items: flex-end;
-
-  margin-bottom: 40px;
 `;
 
 export const GameInfo = styled.div`
@@ -44,5 +44,4 @@ export const Rating = styled.div`
 
 export const CardsContainer = styled.div`
   width: 72vw;
-  padding-left: 24px;
 `;
