+++ src/components/MainBanner/index.tsx
@@ -1,35 +1,32 @@
-import { FaPlaystation, FaSteam, FaXbox } from 'react-icons/fa';
-import { Container, Content, PlatformsContainer, Button } from './styles';
 import '../../utils/swiper-settings';
+import { Swiper, SwiperSlide } from 'swiper/react';
+import MainBannerItem from './MainBannerItem';
 
-interface IMainBannerProps {
-  title: string;
-  summary: string;
-  coverImage: string;
-}
+import data from '../../database/mock';
+import { MainBannerContainer } from './styles';
 
-const MainBanner = ({ title, summary, coverImage }: IMainBannerProps) => {
+const MainBanner = () => {
   return (
-    <Container
-      style={{
-        backgroundSize: 'cover',
-        backgroundImage: `linear-gradient(51deg,rgba(0, 0, 0, 1) 27%,rgba(0, 0, 0, 1) 38%,rgba(0, 0, 0, 0.22452731092436973) 100%), url(${coverImage})`,
-      }}
-    >
-      <Content>
-        <h1>{title}</h1>
-        <p>{summary}</p>
-
-        <PlatformsContainer>
-          <h3>Platforms: </h3>
-          <FaPlaystation size={18} />
-          <FaSteam size={18} />
-          <FaXbox size={18} />
-        </PlatformsContainer>
-
-        <Button type="button">See More</Button>
-      </Content>
-    </Container>
+    <MainBannerContainer>
+      <Swiper
+        spaceBetween={800}
+        autoplay={{
+          delay: 2500,
+        }}
+        parallax
+        speed={600}
+      >
+        {data.map(({ title, summary, coverImage }) => (
+          <SwiperSlide>
+            <MainBannerItem
+              title={title}
+              summary={summary}
+              coverImage={coverImage}
+            />
+          </SwiperSlide>
+        ))}
+      </Swiper>
+    </MainBannerContainer>
   );
 };
