+++ src/pages/Home/index.tsx
@@ -1,41 +1,19 @@
 import React from 'react';
-import { Swiper, SwiperSlide } from 'swiper/react';
 import LargeCard from '../../components/Cards/LargeCard';
 import MediumCard from '../../components/Cards/MediumCard';
 import Header from '../../components/Header';
 import MainBanner from '../../components/MainBanner';
-import data from '../../database/mock';
 
 import '../../utils/swiper-settings';
-import { Container, MainBannerContainer } from './styles';
+import Container from './styles';
 
 const Home: React.FC = () => {
   return (
     <Container>
       <Header />
+      <MainBanner />
 
-      <MainBannerContainer>
-        <Swiper
-          spaceBetween={800}
-          autoplay={{
-            delay: 2500,
-          }}
-          parallax
-          speed={600}
-        >
-          {data.map(({ title, summary, coverImage }) => (
-            <SwiperSlide>
-              <MainBanner
-                title={title}
-                summary={summary}
-                coverImage={coverImage}
-              />
-            </SwiperSlide>
-          ))}
-        </Swiper>
-      </MainBannerContainer>
-
-      <h2>Category</h2>
+      <MediumCard />
       <MediumCard />
 
       <LargeCard />
