+++ src/pages/Home/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const Container = styled.div`
+const Container = styled.div`
   width: 100%;
   .swiper-button-prev,
   .swiper-button-next {
@@ -8,6 +8,4 @@ export const Container = styled.div`
   }
 `;
 
-export const MainBannerContainer = styled.div`
-  width: 72vw;
-`;
+export default Container;
